Fix data exfil even with disabled Sentry logging
Motivation: The `sentry_sdk` package implements pretty
aggressive instrumentation that hooks itself into various
parts of the Python runtime which can result in data sent
to Sentry that the user may not intend to share.

See openclimatefix/open-source-quartz-solar-forecast/317

Ideally, the Sentry logging would be fixed to only do the
explicit logging that the original commit intended, but for
now this change just fixes the handling of the
`QUARTZ_SOLAR_FORECAST_LOGGING` environment variable to at
least fully disable the `sentry_sdk` integration.